refactor: restructure parse_response module with strongly typed output
Reorganize parse_response/ to per-provider layout matching build_request/.
Merge separate _types.rs files into their provider modules. Drop dynamic
metadata map from LlmProviderResponse. Add cached_input_tokens to TokenUsage.
Remove ApiError variant (HTTP errors handled before parsing). Wire OpenAI
and Anthropic parsers to populate cached token counts.
