+++ baml_language/crates/sys_llm/src/parse_response/anthropic.rs
@@ -1,8 +1,51 @@
-use super::{
-    ParseResponseError,
-    anthropic_types::{AnthropicMessageContent, AnthropicMessageResponse},
-    types::{FinishReason, LlmProviderResponse, TokenUsage},
-};
+use serde::Deserialize;
+
+use super::{FinishReason, LlmProviderResponse, ParseResponseError, TokenUsage};
+
+// ── Serde types ───────────────────────────────────────────────────
+
+#[derive(Clone, Debug, Deserialize, PartialEq)]
+#[serde(tag = "type", rename_all = "snake_case")]
+enum AnthropicMessageContent {
+    Text {
+        text: String,
+    },
+    ToolUse {
+        id: Option<String>,
+        input: serde_json::Value,
+        name: String,
+    },
+    RedactedThinking {
+        data: String,
+    },
+    #[serde(other)]
+    Other,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq)]
+#[allow(clippy::struct_field_names)]
+struct AnthropicUsage {
+    pub input_tokens: u64,
+    pub output_tokens: u64,
+    #[serde(default)]
+    pub cache_creation_input_tokens: Option<u64>,
+    #[serde(default)]
+    pub cache_read_input_tokens: Option<u64>,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq)]
+struct AnthropicMessageResponse {
+    pub id: String,
+    pub role: String,
+    pub r#type: String,
+    pub content: Vec<AnthropicMessageContent>,
+    pub model: String,
+    pub stop_reason: Option<String>,
+    pub stop_sequence: Option<serde_json::Value>,
+    pub usage: AnthropicUsage,
+}
+
+// ── Parser ────────────────────────────────────────────────────────
 
 /// Parse an Anthropic message response body into a normalized `LlmProviderResponse`.
 pub(super) fn parse_anthropic_response(
@@ -61,35 +104,8 @@ pub(super) fn parse_anthropic_response(
         input_tokens: Some(input),
         output_tokens: Some(output),
         total_tokens: Some(input + output),
+        cached_input_tokens: response.usage.cache_read_input_tokens,
     };
-
-    // Build metadata map with provider-specific fields.
-    let mut metadata = serde_json::Map::new();
-    metadata.insert("id".into(), serde_json::Value::String(response.id.clone()));
-    if let Some(stop_seq) = &response.stop_sequence {
-        metadata.insert(
-            "stop_sequence".into(),
-            serde_json::Value::String(stop_seq.value.clone()),
-        );
-    }
-    if response.usage.cache_creation_input_tokens > 0 {
-        metadata.insert(
-            "cache_creation_input_tokens".into(),
-            serde_json::Value::Number(response.usage.cache_creation_input_tokens.into()),
-        );
-    }
-    if response.usage.cache_read_input_tokens > 0 {
-        metadata.insert(
-            "cache_read_input_tokens".into(),
-            serde_json::Value::Number(response.usage.cache_read_input_tokens.into()),
-        );
-    }
-    if !response.usage.service_tier.is_empty() {
-        metadata.insert(
-            "service_tier".into(),
-            serde_json::Value::String(response.usage.service_tier.clone()),
-        );
-    }
 
     Ok(LlmProviderResponse {
         content,
@@ -97,7 +113,6 @@ pub(super) fn parse_anthropic_response(
         finish_reason,
         finish_reason_raw: response.stop_reason,
         usage,
-        metadata,
     })
 }
 
@@ -138,12 +153,7 @@ mod tests {
         assert_eq!(resp.usage.input_tokens, Some(9));
         assert_eq!(resp.usage.output_tokens, Some(8));
         assert_eq!(resp.usage.total_tokens, Some(17));
-
-        // Metadata
-        assert_eq!(resp.metadata["id"], "msg_013QyXSmCitiepWfcCMHPTsQ");
-        assert_eq!(resp.metadata["cache_creation_input_tokens"], 51);
-        assert_eq!(resp.metadata["cache_read_input_tokens"], 2258);
-        assert_eq!(resp.metadata["service_tier"], "standard");
+        assert_eq!(resp.usage.cached_input_tokens, Some(2258));
     }
 
     #[test]
@@ -169,7 +179,6 @@ mod tests {
 
         let resp = parse_anthropic_response(body).unwrap();
         assert_eq!(resp.finish_reason, FinishReason::Stop);
-        assert_eq!(resp.metadata["stop_sequence"], "END");
     }
 
     #[test]
@@ -280,7 +289,7 @@ mod tests {
     }
 
     #[test]
-    fn test_cache_tokens_zero_not_in_metadata() {
+    fn test_parse_cached_tokens() {
         let body = r#"{
             "id": "msg_test",
             "type": "message",
@@ -290,16 +299,30 @@ mod tests {
             "stop_reason": "end_turn",
             "stop_sequence": null,
             "usage": {
-                "input_tokens": 5,
-                "output_tokens": 2,
-                "cache_creation_input_tokens": 0,
-                "cache_read_input_tokens": 0
+                "input_tokens": 100,
+                "output_tokens": 10,
+                "cache_read_input_tokens": 50
             }
         }"#;
 
         let resp = parse_anthropic_response(body).unwrap();
-        // Zero cache tokens should not appear in metadata
-        assert!(!resp.metadata.contains_key("cache_creation_input_tokens"));
-        assert!(!resp.metadata.contains_key("cache_read_input_tokens"));
+        assert_eq!(resp.usage.cached_input_tokens, Some(50));
+    }
+
+    #[test]
+    fn test_parse_no_cached_tokens() {
+        let body = r#"{
+            "id": "msg_test",
+            "type": "message",
+            "role": "assistant",
+            "model": "claude-3-haiku-20240307",
+            "content": [{"type": "text", "text": "hi"}],
+            "stop_reason": "end_turn",
+            "stop_sequence": null,
+            "usage": { "input_tokens": 100, "output_tokens": 10 }
+        }"#;
+
+        let resp = parse_anthropic_response(body).unwrap();
+        assert_eq!(resp.usage.cached_input_tokens, None);
     }
 }
