+++ baml_language/crates/sys_llm/src/parse_response/mod.rs
@@ -1,10 +1,5 @@
 mod anthropic;
-mod anthropic_types;
 mod openai;
-mod openai_types;
-mod types;
-
-pub(super) use types::LlmProviderResponse;
 
 use crate::LlmProvider;
 
@@ -16,14 +11,6 @@ pub(crate) enum ParseResponseError {
         #[source]
         source: serde_json::Error,
         content: String,
-    },
-
-    #[error("API error from {provider}: {message}. Code: {code:?}. Param: {param:?}")]
-    ApiError {
-        provider: &'static str,
-        message: String,
-        code: Option<String>,
-        param: Option<String>,
     },
 
     #[error("{provider} response has no content: {detail}")]
@@ -42,11 +29,55 @@ pub(crate) enum ParseResponseError {
     UnsupportedProvider(String),
 }
 
+// ── Shared types ──────────────────────────────────────────────────
+
+// Allows dead_code while consumer only reads content + finish_reason_raw
+#[allow(dead_code)]
+/// Normalized response from any LLM provider.
+#[derive(Debug, Clone)]
+pub(crate) struct LlmProviderResponse {
+    /// Text content extracted from the LLM response.
+    pub content: String,
+    /// Model identifier returned by the provider.
+    pub model: String,
+    /// Normalized finish reason.
+    pub finish_reason: FinishReason,
+    /// Raw provider-reported finish reason string.
+    pub finish_reason_raw: Option<String>,
+    /// Token usage information.
+    pub usage: TokenUsage,
+}
+
+/// Normalized finish reason.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) enum FinishReason {
+    Stop,
+    Length,
+    ToolUse,
+    Other(String),
+    Unknown,
+}
+
+#[allow(dead_code)]
+impl FinishReason {
+    pub(crate) fn is_complete(&self) -> bool {
+        matches!(self, FinishReason::Stop)
+    }
+}
+
+/// Token usage reported by the provider.
+#[allow(clippy::struct_field_names, dead_code)]
+#[derive(Debug, Clone, Default)]
+pub(crate) struct TokenUsage {
+    pub input_tokens: Option<u64>,
+    pub output_tokens: Option<u64>,
+    pub total_tokens: Option<u64>,
+    pub cached_input_tokens: Option<u64>,
+}
+
+// ── Dispatcher ────────────────────────────────────────────────────
+
 /// Parse a raw HTTP response body into a normalized `LlmProviderResponse`.
-///
-/// The `provider` determines which deserialization format to use. Callers
-/// should resolve the effective provider before calling (e.g. Vertex AI +
-/// Anthropic model should pass `LlmProvider::Anthropic`).
 pub(crate) fn parse_response(
     provider: LlmProvider,
     body: &str,
@@ -56,7 +87,7 @@ pub(crate) fn parse_response(
         | LlmProvider::OpenAiGeneric
         | LlmProvider::AzureOpenAi
         | LlmProvider::Ollama
-        | LlmProvider::OpenRouter => openai::parse_openai_response(body),
+        | LlmProvider::OpenRouter => openai::chat_completions::parse_openai_response(body),
 
         LlmProvider::Anthropic | LlmProvider::AwsBedrock => {
             anthropic::parse_anthropic_response(body)
